requirements.txt
langgraph>=1.0.0
langchain-openai>=0.3.0
langchain-core>=1.0.0
duckduckgo-search>=6.0
pydantic>=2.0
python-dotenv>=1.0
PyJWT>=2.0

# Anthropic
# langchain-anthropic>=0.3.0

# Ollama
# langchain-community>=0.2

# Database (PostgreSQL — required)
sqlalchemy>=2.0
psycopg2-binary>=2.9        # SQLAlchemy driver
psycopg[binary]>=3.1        # LangGraph PostgresSaver driver

# Migrations
alembic>=1.13

# LangGraph checkpoint (PostgreSQL)
langgraph-checkpoint-postgres>=2.0

# Web scraping (scrape_page tool)
requests>=2.31
beautifulsoup4>=4.12         # optional: cleaner HTML → text (falls back to regex)

# FastAPI + ASGI server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9     # form data support

# MCP — Model Context Protocol plugin system (optional, gracefully skipped if absent)
langchain-mcp-adapters>=0.1.0

# Redis + async worker queue (Phase 1 — Scalability)
arq>=0.26
redis[hiredis]>=5.0

# Testing
pytest>=7.0
pytest-mock>=3.0
pytest-asyncio>=0.21
